Make the code -safe-string compatible
We now depend on the 'bytes' ocamlfind package, which provides
compatibility of Bytes-using code on older OCaml versions.

The only place where the new code is not equivalent to the previous
one is the reading of section names in index.ml: we allocate a new
string for each section name read, while the previous code mutated
a buffer in place. Allocating a Bytes buffer in place would be doable,
but then we would use Bytes.to_string to match against section name
strings (so doing an allocation in any case), or would have to use an
unsafe operation in a place where that does not feel necessary.